# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RiskSense AI API",
    "version": "0.1.0",
    "description": "Generated from Zod schemas. See docs/ai/API.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "nullable": true
              }
            },
            "required": [
              "code",
              "message"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "requestId"
            ]
          }
        },
        "required": [
          "error",
          "meta"
        ]
      },
...
      "Question": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        }
      }
    },
...
  },
  "paths": {
...
    "/questions": {
      "get": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "personaKey",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "scenarioKey",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "financial",
                "healthcare",
                "it",
                "general"
              ]
            },
            "required": false,
            "name": "sector",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "retired"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List Questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Question"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,63}$"
                  },
                  "text": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "mcq",
                      "yes_no",
                      "free_text",
                      "number"
                    ]
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "label": {
                          "type": "string",
                          "minLength": 1
                        },
                        "factValue": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "label",
                        "factValue"
                      ]
                    },
                    "default": []
                  },
                  "factKey": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,63}$"
                  },
                  "required": {
                    "type": "boolean",
                    "default": true
                  },
                  "tags": {
                    "type": "object",
                    "properties": {
                      "personaKeys": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "minItems": 1
                      },
                      "scenarioKeys": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "default": []
                      },
                      "sectors": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "financial",
                            "healthcare",
                            "it",
                            "general"
                          ]
                        },
                        "minItems": 1
                      },
                      "category": {
                        "type": "string",
                        "maxLength": 80
                      }
                    },
                    "required": [
                      "personaKeys",
                      "sectors"
                    ],
                    "additionalProperties": false
                  },
                  "branchTrigger": {
                    "type": "object",
                    "properties": {
                      "onValue": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          }
                        ]
                      },
                      "questionKeys": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "minItems": 1
                      }
                    },
                    "required": [
                      "onValue",
                      "questionKeys"
                    ]
                  },
                  "scoringHint": {
                    "type": "string",
                    "maxLength": 300
                  }
                },
                "required": [
                  "key",
                  "text",
                  "type",
                  "factKey",
                  "tags"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Question created (draft)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Question"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT (duplicate key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/questions/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Question",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Question"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "mcq",
                      "yes_no",
                      "free_text",
                      "number"
                    ]
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "label": {
                          "type": "string",
                          "minLength": 1
                        },
                        "factValue": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "label",
                        "factValue"
                      ]
                    },
                    "default": []
                  },
                  "factKey": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,63}$"
                  },
                  "required": {
                    "type": "boolean",
                    "default": true
                  },
                  "tags": {
                    "type": "object",
                    "properties": {
                      "personaKeys": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "minItems": 1
                      },
                      "scenarioKeys": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "default": []
                      },
                      "sectors": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "financial",
                            "healthcare",
                            "it",
                            "general"
                          ]
                        },
                        "minItems": 1
                      },
                      "category": {
                        "type": "string",
                        "maxLength": 80
                      }
                    },
                    "required": [
                      "personaKeys",
                      "sectors"
                    ],
                    "additionalProperties": false
                  },
                  "branchTrigger": {
                    "type": "object",
                    "properties": {
                      "onValue": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          }
                        ]
                      },
                      "questionKeys": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "minItems": 1
                      }
                    },
                    "required": [
                      "onValue",
                      "questionKeys"
                    ]
                  },
                  "scoringHint": {
                    "type": "string",
                    "maxLength": 300
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Question"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/questions/{id}/retire": {
      "post": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Retired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Question"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}